100/100

+++ j10/ex06/calculator.h
@@ -6,24 +6,24 @@
 /*   By: alcaroff <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/24 14:00:40 by alcaroff          #+#    #+#             */
-/*   Updated: 2017/07/26 09:51:47 by alcaroff         ###   ########.fr       */
+/*   Updated: 2017/07/25 11:41:54 by alcaroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef CALCULATOR_H
 # define CALCULATOR_H
 
-long			ft_add(long a, long b);
-long			ft_sub(long a, long b);
-long			ft_mul(long a, long b);
-long			ft_div(long a, long b);
-long			ft_mod(long a, long b);
-long			ft_usage(long a, long b);
+int				ft_add(int a, int b);
+int				ft_sub(int a, int b);
+int				ft_mul(int a, int b);
+int				ft_div(int a, int b);
+int				ft_mod(int a, int b);
+int				ft_usage(int a, int b);
 
 typedef struct	s_opp
 {
 	char	*op;
-	long	(*fun)(long, long);
+	int		(*fun)(int, int);
 }				t_opp;
 
 #endif
